requirements.txt
globus-sdk
globus-cli
globus-compute-endpoint
numpy
matplotlib
python-dotenv
python-dateutil
argparse
shortuuid